giza: refactoring main sphinx entry point
